+++ libraries/apollo-mockserver/api/apollo-mockserver.api
@@ -1,6 +1,3 @@
-public final class com/apollographql/apollo3/mockserver/-MockServers {
-}
-
 public final class com/apollographql/apollo3/mockserver/MockRequest {
 	public fun <init> (Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;Ljava/util/Map;Lokio/ByteString;)V
 	public synthetic fun <init> (Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;Ljava/util/Map;Lokio/ByteString;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
@@ -42,12 +39,6 @@ public final class com/apollographql/apollo3/mockserver/MockServer : com/apollog
 	public fun url (Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
 }
 
-public final class com/apollographql/apollo3/mockserver/MockServerCommonKt {
-	public static final fun parseHeader (Ljava/lang/String;)Lkotlin/Pair;
-	public static final fun parseRequestLine (Ljava/lang/String;)Lkotlin/Triple;
-	public static final fun writeResponse (Lokio/BufferedSink;Lcom/apollographql/apollo3/mockserver/MockResponse;Ljava/lang/String;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
-}
-
 public abstract interface class com/apollographql/apollo3/mockserver/MockServerHandler {
 	public abstract fun handle (Lcom/apollographql/apollo3/mockserver/MockRequest;)Lcom/apollographql/apollo3/mockserver/MockResponse;
 }
@@ -58,6 +49,9 @@ public abstract interface class com/apollographql/apollo3/mockserver/MockServerI
 	public abstract fun stop (Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
 	public abstract fun takeRequest ()Lcom/apollographql/apollo3/mockserver/MockRequest;
 	public abstract fun url (Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
+}
+
+public final class com/apollographql/apollo3/mockserver/MockServerKt {
 }
 
 public final class com/apollographql/apollo3/mockserver/WebSocketMockServer_jvmKt {
