+++ libraries/apollo-mockserver/src/commonMain/kotlin/com/apollographql/apollo3/mockserver/MockRequest.kt
@@ -1,0 +1,11 @@
+package com.apollographql.apollo3.mockserver
+
+import okio.ByteString
+
+class MockRequest(
+    val method: String,
+    val path: String,
+    val version: String,
+    val headers: Map<String, String> = emptyMap(),
+    val body: ByteString = ByteString.EMPTY,
+)
