+++ libraries/apollo-mockserver/src/commonMain/kotlin/com/apollographql/apollo3/mockserver/MockServerHandler.kt
@@ -1,0 +1,10 @@
+package com.apollographql.apollo3.mockserver
+
+interface MockServerHandler {
+  /**
+   * Handles the given [MockRequest].
+   *
+   * This method is called from one or several background threads and must be thread-safe.
+   */
+  fun handle(request: MockRequest): MockResponse
+}
