+++ libraries/apollo-mockserver/src/commonTest/kotlin/com/apollographql/apollo3/mockserver/test/EnqueueTest.kt
@@ -1,5 +1,6 @@
 package com.apollographql.apollo3.mockserver.test
 
+import com.apollographql.apollo3.annotations.ApolloInternal
 import com.apollographql.apollo3.api.http.HttpMethod
 import com.apollographql.apollo3.api.http.HttpRequest
 import com.apollographql.apollo3.mockserver.MockResponse
@@ -11,6 +12,7 @@ import kotlinx.coroutines.flow.flowOf
 import okio.ByteString.Companion.encodeUtf8
 import kotlin.test.Test
 
+@ApolloInternal
 class EnqueueTest {
   private lateinit var mockServer: MockServer
